Apply Debian patch 01_Makefile.am.patch (part)
Change LIBZVBI_CHAINS_PATH value for zvbi-chains utility to
"libzvbi-chains.so.0" instead of "libzvbi-chains.so". Different from
Debian patch that uses libzvbi-chains.so.0.0.0 specifically. Loading .0
is less constrained but should still work with the Debian libzvbi0
package because it includes both, just not the unversioned .so file.

Original path written on Wed, 20 May 2015.