Replace strptime implementation for win32
There are multiple reasons for this change. The main is that previous
implementation was not working at all. Not only it crashes when parsing
valid input, but also fails on %H%M%S formats, which are the main we
care about. It would parse 123456 as one number, because there was no
max lenght, which should be 2 in this case. Above that it was complex
and big, for no good reason and still not working as expected.

Instead of fixing multiple issues with it, replace it with small and
simple but functional implementation borrowed from FFmpeg. It is good
enough to pass all our tests.

Fixes data string parsing.

Implementation copied from FFmpeg.